requirements.txt
# Core Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# PyTorch (required for albumentations)
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
opencv-python-headless>=4.8.0

# Data Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics & Evaluation
scikit-learn>=1.3.0

# Progress Bar
tqdm>=4.65.0

# Utilities
scipy>=1.10.0
pyyaml>=6.0

# Jupyter (optional - for notebooks)
# jupyter>=1.0.0
# ipykernel>=6.23.0

# Experiment Tracking (optional)
# wandb>=0.15.0

# Deployment (optional)
streamlit>=1.25.0
# flask>=2.3.0
# fastapi>=0.100.0
